refactor : remove dependency of LottoAnswer in Lotto
LottoAnswer extends Lotto, so Lotto should not have dependency of LottoAnswer, which it's child class
Lotto
LottoAnswer override compareLotto to remove dependency of LottoAnswer in Lotto
