+++ src/test/java/lotto/model/domain/LottoTest.java
@@ -1,19 +1,17 @@
 package lotto.model.domain;
 
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import java.util.List;
 import java.util.stream.Stream;
 import lotto.model.domain.lotto.Lotto;
-import lotto.model.domain.lotto.LottoAnswer;
 import lotto.model.domain.result.LottoCompareResult;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class LottoTest {
     @DisplayName("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.")
@@ -74,34 +72,4 @@ class LottoTest {
         );
     }
 
-    @ParameterizedTest
-    @MethodSource("equalsWithAnswerArgs")
-    void 정답과_동등성_테스트(Lotto l1, Lotto l2, LottoCompareResult result) {
-        Assertions.assertThat(l1.compareLotto(l2))
-                .isEqualTo(result);
-    }
-
-    static Stream<Arguments> equalsWithAnswerArgs() {
-        return Stream.of(
-                Arguments.of(new Lotto(List.of(1, 2, 3, 4, 5, 6)),
-                        new LottoAnswer(List.of(1, 2, 3, 4, 5, 6), 9),
-                        new LottoCompareResult(6, false)),
-
-                Arguments.of(new Lotto(List.of(1, 2, 3, 4, 5, 6)),
-                        new LottoAnswer(List.of(2, 3, 4, 5, 6, 7), 10),
-                        new LottoCompareResult(5, false)),
-
-                Arguments.of(new Lotto(List.of(1, 2, 3, 4, 5, 6)),
-                        new LottoAnswer(List.of(2, 3, 4, 5, 6, 7), 1),
-                        new LottoCompareResult(5, true)),
-
-                Arguments.of(new Lotto(List.of(1, 2, 3, 4, 5, 6)),
-                        new LottoAnswer(List.of(7, 8, 9, 10, 11, 12), 6),
-                        new LottoCompareResult(0, false)),
-
-                Arguments.of(new Lotto(List.of(1, 3, 6, 15, 19, 31)),
-                        new LottoAnswer(List.of(6, 11, 19, 23, 31, 42), 1),
-                        new LottoCompareResult(3, false))
-        );
-    }
 }
